refactor (BpostApiResponseException): Rename from ApiResponseException
