--- src/Exception/ApiResponseException/BpostApiBusinessException.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Bpost\BpostApiClient\Exception\ApiResponseException;
-
-use Bpost\BpostApiClient\Exception\ApiResponseException;
-
-/**
- * Class BpostApiBusinessException
- * @package Bpost\BpostApiClient\Exception\ApiResponseException
- */
-class BpostApiBusinessException extends ApiResponseException
-{
-}
