--- src/Exception/ApiResponseException/BpostApiValidationException.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Bpost\BpostApiClient\Exception\ApiResponseException;
-
-use Bpost\BpostApiClient\Exception\ApiResponseException;
-
-/**
- * Class BpostApiValidationException
- * @package Bpost\BpostApiClient\Exception\ApiResponseException
- */
-class BpostApiValidationException extends ApiResponseException
-{
-}
