--- src/Exception/ApiResponseException/BpostCurlException.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Bpost\BpostApiClient\Exception\ApiResponseException;
-
-use Bpost\BpostApiClient\Exception\ApiResponseException;
-
-/**
- * Class BpostCurlException
- * @package Bpost\BpostApiClient\Exception\ApiResponseException
- */
-class BpostCurlException extends ApiResponseException
-{
-}
