--- src/Exception/ApiResponseException/BpostInvalidSelectionException.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Bpost\BpostApiClient\Exception\ApiResponseException;
-
-use Bpost\BpostApiClient\Exception\ApiResponseException;
-
-/**
- * Class BpostInvalidSelectionException
- * @package Bpost\BpostApiClient\Exception\ApiResponseException
- */
-class BpostInvalidSelectionException extends ApiResponseException
-{
-}
