--- src/Exception/ApiResponseException/BpostTaxipostLocatorException.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Bpost\BpostApiClient\Exception\ApiResponseException;
-
-use Bpost\BpostApiClient\Exception\ApiResponseException;
-
-/**
- * Class BpostTaxipostLocatorException
- * @package Bpost\BpostApiClient\Exception\ApiResponseException
- */
-class BpostTaxipostLocatorException extends ApiResponseException
-{
-}
