+++ src/Exception/BpostApiResponseException/BpostApiBusinessException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+use Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+/**
+ * Class BpostApiBusinessException
+ * @package Bpost\BpostApiClient\Exception\BpostApiResponseException
+ */
+class BpostApiBusinessException extends BpostApiResponseException
+{
+}
