+++ src/Exception/BpostApiResponseException/BpostApiValidationException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+use Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+/**
+ * Class BpostApiValidationException
+ * @package Bpost\BpostApiClient\Exception\BpostApiResponseException
+ */
+class BpostApiValidationException extends BpostApiResponseException
+{
+}
