+++ src/Exception/BpostApiResponseException/BpostCurlException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+use Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+/**
+ * Class BpostCurlException
+ * @package Bpost\BpostApiClient\Exception\BpostApiResponseException
+ */
+class BpostCurlException extends BpostApiResponseException
+{
+}
