+++ src/Exception/BpostApiResponseException/BpostInvalidSelectionException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+use Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+/**
+ * Class BpostInvalidSelectionException
+ * @package Bpost\BpostApiClient\Exception\BpostApiResponseException
+ */
+class BpostInvalidSelectionException extends BpostApiResponseException
+{
+}
