+++ src/Exception/BpostApiResponseException/BpostTaxipostLocatorException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+use Bpost\BpostApiClient\Exception\BpostApiResponseException;
+
+/**
+ * Class BpostTaxipostLocatorException
+ * @package Bpost\BpostApiClient\Exception\BpostApiResponseException
+ */
+class BpostTaxipostLocatorException extends BpostApiResponseException
+{
+}
