+++ tests/Exception/ApiResponseException/BpostApiBusinessExceptionTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Bpost\BpostApiClient\test\Exception\ApiResponseException;
+namespace Bpost\BpostApiClient\test\Exception\BpostApiResponseException;
 
-use Bpost\BpostApiClient\Exception\ApiResponseException\BpostApiBusinessException;
+use Bpost\BpostApiClient\Exception\BpostApiResponseException\BpostApiBusinessException;
 
 class BpostApiBusinessExceptionTest extends \PHPUnit_Framework_TestCase
 {
