+++ tests/Exception/ApiResponseException/BpostTaxipostLocatorExceptionTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Bpost\BpostApiClient\test\Exception\ApiResponseException;
+namespace Bpost\BpostApiClient\test\Exception\BpostApiResponseException;
 
-use Bpost\BpostApiClient\Exception\ApiResponseException\BpostTaxipostLocatorException;
+use Bpost\BpostApiClient\Exception\BpostApiResponseException\BpostTaxipostLocatorException;
 
 class BpostTaxipostLocatorExceptionTest extends \PHPUnit_Framework_TestCase
 {
